Refactored min/max search
