+++ metrics_collector.py
@@ -1,12 +1,12 @@
 import asyncio
 import time
 
-from VolatilityTask.base_volatility_integration import StockMarketIntegration
-from VolatilityTask.binance_volatility_integration import BinanceStockMarketIntegration
+from VolatilityTask.integration.base_integration import BaseStockMarketIntegration
+from VolatilityTask.integration.binance_integration import BinanceBaseStockMarketIntegration
 from VolatilityTask.consts import SYMBOL_PAIRS
 
 
-async def collect_metrics(integration: StockMarketIntegration):
+async def collect_metrics(integration: BaseStockMarketIntegration):
     result = await asyncio.create_task(integration.calculate_all_symbols_metrics_for_last_hour())
     for pair_metrics in result:
         print(pair_metrics)
@@ -14,6 +14,6 @@ async def collect_metrics(integration: StockMarketIntegration):
 
 start_time = time.time()
 
-asyncio.run(collect_metrics(BinanceStockMarketIntegration(SYMBOL_PAIRS)))
+asyncio.run(collect_metrics(BinanceBaseStockMarketIntegration(SYMBOL_PAIRS)))
 
 print((time.time() - start_time))
